third_party/WebKit/LayoutTests/fast/canvas/canvas-hit-regions-fill-rule-test-expected.txt

default fillRule(nonzero):
PASS clickCanvas(1, 1) is 'nonzero'
PASS clickCanvas(99, 98) is 'nonzero'
PASS clickCanvas(21, 21) is 'nonzero'
PASS clickCanvas(50, 50) is 'nonzero'

fillRule = evenodd:
PASS clickCanvas(1, 1) is 'evenodd'
PASS clickCanvas(99, 98) is 'evenodd'
PASS clickCanvas(21, 21) is null
PASS clickCanvas(50, 50) is null

fillRule = nonzero:
PASS clickCanvas(1, 1) is 'nonzero'
PASS clickCanvas(99, 98) is 'nonzero'
PASS clickCanvas(21, 21) is 'nonzero'
PASS clickCanvas(50, 50) is 'nonzero'

fillRule = null, falls back to default (nonzero):
PASS clickCanvas(1, 1) is 'nonzero'
PASS clickCanvas(99, 98) is 'nonzero'
PASS clickCanvas(21, 21) is 'nonzero'
PASS clickCanvas(50, 50) is 'nonzero'

fillRule = undefined, falls back to default (nonzero):
PASS clickCanvas(1, 1) is 'nonzero'
PASS clickCanvas(99, 98) is 'nonzero'
PASS clickCanvas(21, 21) is 'nonzero'
PASS clickCanvas(50, 50) is 'nonzero'

PASS successfullyParsed is true

TEST COMPLETE
